Implement 'deep clean' which removes third party sources.

+++ Build/source/clean.js
@@ -13,12 +13,17 @@
 // limitations under the License.
 
 const fs = require('fs');
+const path = require('path');
 const { forgetAlreadyBuiltLibraries } = require('./build');
 const { getTempDirectory, getOutputPath, getFileSystemDirectory } = require('./config');
 const { forgetAllLastModifiedTimestamps } = require('./file_timestamps');
+const { getAllApplications, getPackageDirectory, getAllLibraries } = require('./package_directory');
+const { PackageType } = require('./package_type');
+const { removeDirectoryIfEmpty } = require('./utils');
 
 // Tries to delete a file or directory, if it exists.
 function maybeDelete(path) {
+  console.log(path);
   if (!fs.existsSync(path)) {
     return;
   }
@@ -30,13 +35,58 @@ function maybeDelete(path) {
   }
 }
 
+function maybeDeleteAndRemoveParentIfEmpty(filePath) {
+  maybeDelete(filePath);
+  removeDirectoryIfEmpty(path.dirName(filePath));
+}
+
+function deepCleanPackage(packageDir) {
+  const thirdPartyFilesPath = packageDir + 'third_party_files.json';
+  if (fs.existsSync(thirdPartyFilesPath)) {
+    try {
+      const thirdPartyFiles = JSON.parse(fs.readFileSync(thirdPartyFilesPath));
+      Object.keys(thirdPartyFiles).forEach(maybeDeleteAndRemoveParentIfEmpty);
+      //maybeDelete(thirdPartyFilesPath);
+    } catch {
+      console.log('Error deleting third party files in ' + packageDir);
+    }
+  }
+
+  maybeDelete(packageDir + 'generated');
+}
+
 // Removes all built files.
-function clean() {
+function clean(deepClean) {
   forgetAlreadyBuiltLibraries();
   forgetAllLastModifiedTimestamps();
 
-  if (fs.existsSync(getTempDirectory()))
-    fs.rmSync(getTempDirectory(), { recursive: true });
+  if (deepClean) {
+    // Delete all temp files including repositories.
+    if (fs.existsSync(getTempDirectory())) {
+      fs.rmSync(getTempDirectory(), { recursive: true });
+    }
+
+    // Delete all generated and third party files.
+    getAllApplications().forEach(application => {
+      deepCleanPackage(getPackageDirectory(PackageType.APPLICATION, application));
+    });
+
+    getAllLibraries().forEach(library => {
+      deepCleanPackage(getPackageDirectory(PackageType.LIBRARY, library));
+    });
+
+  } else {
+    // Don't delete repositories.
+    if (fs.existsSync(getTempDirectory())) {
+      let tempEntries = fs.readdirSync(getTempDirectory());
+      tempEntries.forEach(tempEntry => {
+        if (tempEntry == 'repositories')
+          return;
+
+        maybeDelete(getTempDirectory() + '/' + tempEntry);
+      });
+    }
+  }
 
   // Clean up FS image.
   if (fs.existsSync(getFileSystemDirectory() + '/Applications'))
